[feature/19-create-employer-endpoint]  (#21)
* [feature/19-create-employer-endpoint] employer endpoint added, employees added into cafe model. Delete user added into adminController.

* [feature/10-create-employer-endpoint] employer and admin controllers added

* [feature/19-create-employer-endpoint] endpoints are updated. swagger updated. postman ready to share.

* [feature/19-create-employer-endpoint] review yorumlarına göre düzenlendi

---------

+++ controllers/commentController.js
@@ -9,7 +9,7 @@ exports.getCommentsByCafeId = catchAsyncErrors(async (req, res, next) => {
     if (!cafe) {
         return next(new ErrorHandler('Cafe not found', 404))
     }
-    const comments = await Comment.find({ cafe: cafe })
+    const comments = await Comment.find({ cafe: cafe }).populate('user')
     res.status(200).json({
         success: true,
         message: null,
@@ -24,8 +24,12 @@ exports.createComment = catchAsyncErrors(async (req, res, next) => {
         return next(new ErrorHandler('Cafe not found', 404))
     }
 
-    if (cafe.owner == req.user.id || cafe.employees.includes(req.user.id)) {
-        return next(new ErrorHandler('Cafe employee or employees can not make comment', 400))
+    if (cafe.owner == req.user.id) {
+        return next(new ErrorHandler('Cafe employer can not make comment', 400))
+    }
+
+    if (cafe.employees.some(employee => employee.user == req.user.id)) {
+        return next(new ErrorHandler('Cafe employee can not make comment', 400))
     }
 
     const { error } = createCommentBodyValidation(req.body)
@@ -47,8 +51,13 @@ exports.createComment = catchAsyncErrors(async (req, res, next) => {
 
 exports.deleteComment = catchAsyncErrors(async (req, res, next) => {
     const comment = await Comment.findById(req.params.commentId)
+    const userCafeId = req.user.cafe
     if (!comment) {
         return next(new ErrorHandler('Comment not found', 404))
+    }
+
+    if (!comment.cafe.equals(userCafeId)) {
+        return next(new ErrorHandler('You have to owner of this cafe to delete this comment', 404))
     }
     await Comment.findByIdAndDelete(req.params.commentId)
     res.status(200).json({
